requirements/base.txt
django<1.12
coreapi==2.3.0
django-autocomplete-light==3.2.1
django-cors-headers==2.0.2
django-extra-fields==0.3
django-filter==1.0.2
django-import-export==0.5.1
django-multiselectfield==0.1.5
django-redis==4.7.0
djangorestframework-jsonapi==2.1.1
djangorestframework<3.8
drfdocs==0.0.11
markdown==2.6.7
psycopg2-binary<2.8
pygments==2.1.3
python-dateutil==2.5.3
requests==2.18.1
